Move reward subtype back into reward condition entity
