+++ src/Contrib/Transformers/MonsterTransformer.php
@@ -11,7 +11,6 @@
 	use App\Entity\RewardCondition;
 	use App\Entity\Strings\MonsterStrings;
 	use App\Entity\Strings\MonsterWeaknessStrings;
-	use App\Entity\Strings\RewardConditionStrings;
 	use App\Localization\L10nUtil;
 	use DaybreakStudios\Utility\DoctrineEntities\EntityInterface;
 	use DaybreakStudios\Utility\EntityTransformers\Exceptions\EntityTransformerException;
@@ -167,14 +166,14 @@
 							);
 						}
 
-						$condition = new RewardCondition(
-							$definition->type,
-							$definition->rank,
-							$definition->quantity,
-							$definition->chance
+						$reward->getConditions()->add(
+							$condition = new RewardCondition(
+								$definition->type,
+								$definition->rank,
+								$definition->quantity,
+								$definition->chance
+							)
 						);
-
-						$reward->getConditions()->add($condition);
 
 						if (ObjectUtil::isset($definition, 'subtype')) {
 							$subtype = $definition->subtype;
@@ -182,7 +181,7 @@
 							if (is_string($subtype))
 								$subtype = strtolower($subtype);
 
-							$this->getRewardConditionStrings($condition)->setSubtype($subtype);
+							$condition->setSubtype($subtype);
 						}
 					}
 				}
@@ -214,18 +213,6 @@
 		protected function getWeaknessStrings(MonsterWeakness $weakness): MonsterWeaknessStrings {
 			$strings = L10nUtil::findOrCreateStrings($this->getCurrentLocale(), $weakness);
 			assert($strings instanceof MonsterWeaknessStrings);
-
-			return $strings;
-		}
-
-		/**
-		 * @param RewardCondition $rewardCondition
-		 *
-		 * @return RewardConditionStrings
-		 */
-		protected function getRewardConditionStrings(RewardCondition $rewardCondition): RewardConditionStrings {
-			$strings = L10nUtil::findOrCreateStrings($this->getCurrentLocale(), $rewardCondition);
-			assert($strings instanceof RewardConditionStrings);
 
 			return $strings;
 		}
