+++ src/Entity/RewardCondition.php
@@ -1,24 +1,19 @@
 <?php
 	namespace App\Entity;
 
-	use App\Entity\Strings\RewardConditionStrings;
 	use App\Game\Rank;
 	use App\Game\RewardConditionType;
-	use App\Localization\TranslatableEntityInterface;
 	use DaybreakStudios\Utility\DoctrineEntities\EntityInterface;
-	use Doctrine\Common\Collections\ArrayCollection;
-	use Doctrine\Common\Collections\Collection;
-	use Doctrine\Common\Collections\Selectable;
 	use Doctrine\ORM\Mapping as ORM;
 	use Symfony\Component\Validator\Constraints as Assert;
 
 	/**
-	 * @ORM\Entity(readOnly=true)
+	 * @ORM\Entity()
 	 * @ORM\Table(name="reward_conditions")
 	 *
 	 * @package App\Entity
 	 */
-	class RewardCondition implements EntityInterface, TranslatableEntityInterface {
+	class RewardCondition implements EntityInterface {
 		use EntityTrait;
 
 		/**
@@ -64,19 +59,13 @@
 		private $chance;
 
 		/**
-		 * @Assert\Valid()
+		 * @Assert\Length(max="128")
 		 *
-		 * @ORM\OneToMany(
-		 *     targetEntity="App\Entity\Strings\RewardConditionStrings",
-		 *     mappedBy="rewardCondition",
-		 *     orphanRemoval=true,
-		 *     cascade={"all"},
-		 *     fetch="EAGER"
-		 * )
+		 * @ORM\Column(type="string", length=128, nullable=true)
 		 *
-		 * @var Collection|Selectable|RewardConditionStrings[]
+		 * @var string|null
 		 */
-		private $strings;
+		private $subtype = null;
 
 		/**
 		 * RewardCondition constructor.
@@ -91,8 +80,6 @@
 			$this->rank = $rank;
 			$this->quantity = $quantity;
 			$this->chance = $chance;
-
-			$this->strings = new ArrayCollection();
 		}
 
 		/**
@@ -168,20 +155,20 @@
 		}
 
 		/**
-		 * @return Collection|Selectable|RewardConditionStrings[]
+		 * @return string|null
 		 */
-		public function getStrings(): Collection {
-			return $this->strings;
+		public function getSubtype(): ?string {
+			return $this->subtype;
 		}
 
 		/**
-		 * @param string $language
+		 * @param string|null $subtype
 		 *
-		 * @return RewardConditionStrings
+		 * @return $this
 		 */
-		public function addStrings(string $language): EntityInterface {
-			$this->getStrings()->add($strings = new RewardConditionStrings($this, $language));
+		public function setSubtype(?string $subtype) {
+			$this->subtype = $subtype;
 
-			return $strings;
+			return $this;
 		}
 	}
